consolidate dashboard contributions; move manage action to dashboard contributions; make groupings the same
